feat: Repackage plugin to align with new GitHub username and update project structure
